+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Fix bug on `Rdmm::Resources::ItemResource#affiliate_url`
 - Fix bug on `Rdmm::Resources::ItemResource#url`
 - Rename `#affiliate_url` to `#url_affiliate` on `Rdmm::Resources::ItemResource`
+- Remove `#prices` on `Rdmm::Resources::ItemResource`
+- Add more methods on `Rdmm::Resources::ItemResource`
 
 ## 0.2.0